refactor: store sold2_detector config dataclasses in utils/structures.py
